api_public/requirements.txt
fastapi>=0.111
uvicorn[standard]
openai>=1.0
numpy
aiosqlite
structlog
python-dotenv
